refactor(agents): extract AgentMcpSessionService from vault service (NV-8053)
Move per-turn vault id and connected MCP resolution into a dedicated session
service so McpConnectionVaultService keeps only vault CRUD for OAuth callbacks.
