+++ apps/api/src/app/agents/mcp/connections/mcp-connection-vault.service.ts
@@ -1,12 +1,7 @@
 import { Injectable } from '@nestjs/common';
 import { type IAgentRuntimeProvider, PinoLogger } from '@novu/application-generic';
-import {
-  AgentMcpServerEntity,
-  AgentMcpServerRepository,
-  McpConnectionEntity,
-  McpConnectionRepository,
-} from '@novu/dal';
-import { MCP_SERVERS, McpConnectionScopeEnum, McpConnectionStatusEnum } from '@novu/shared';
+import { AgentMcpServerRepository, McpConnectionEntity, McpConnectionRepository } from '@novu/dal';
+import { McpConnectionScopeEnum } from '@novu/shared';
 
 @Injectable()
 export class McpConnectionVaultService {
@@ -16,39 +11,6 @@ export class McpConnectionVaultService {
     private readonly logger: PinoLogger
   ) {
     this.logger.setContext(this.constructor.name);
-  }
-
-  /**
-   * Resolve Anthropic `vault_ids` for a managed-agent turn.
-   *
-   * Subscriber-scoped only in v1: anonymous platform turns (no subscriber)
-   * receive `[]` because no write path creates `scope: 'agent'` rows today.
-   * When OAuth MCPs are enabled but no vault exists yet, one is created so
-   * credentials can be bound via `vault_ids` after the subscriber completes
-   * in-thread OAuth (dispatch is gated until then).
-   */
-  async resolveVaultIds(params: {
-    agentId: string;
-    environmentId: string;
-    organizationId: string;
-    subscriberMongoId?: string;
-    runtimeProvider?: IAgentRuntimeProvider;
-  }): Promise<string[]> {
-    if (!params.subscriberMongoId) {
-      return [];
-    }
-
-    const enabledAgentMcpServerIds = await this.listAgentMcpServerIds(params, true);
-    const subscriberVaultId = await this.resolveSubscriberVaultId({
-      agentId: params.agentId,
-      environmentId: params.environmentId,
-      organizationId: params.organizationId,
-      subscriberMongoId: params.subscriberMongoId,
-      agentMcpServerIds: enabledAgentMcpServerIds,
-      runtimeProvider: params.runtimeProvider,
-    });
-
-    return subscriberVaultId ? [subscriberVaultId] : [];
   }
 
   /**
@@ -139,286 +101,6 @@ export class McpConnectionVaultService {
     return externalVaultId;
   }
 
-  private async resolveSubscriberVaultId(params: {
-    agentId: string;
-    environmentId: string;
-    organizationId: string;
-    subscriberMongoId: string;
-    agentMcpServerIds: string[];
-    runtimeProvider?: IAgentRuntimeProvider;
-  }): Promise<string | null> {
-    const existingVaultId = await this.mcpConnectionRepository.findSubscriberExternalVaultId({
-      organizationId: params.organizationId,
-      environmentId: params.environmentId,
-      subscriberId: params.subscriberMongoId,
-      agentMcpServerIds: params.agentMcpServerIds,
-    });
-
-    if (existingVaultId) {
-      return existingVaultId;
-    }
-
-    if (!params.runtimeProvider?.capabilities.tokenVault) {
-      return null;
-    }
-
-    const oauthEnablements = await this.agentMcpServerRepository.findOAuthEnablementsForAgent({
-      organizationId: params.organizationId,
-      environmentId: params.environmentId,
-      agentId: params.agentId,
-    });
-
-    if (oauthEnablements.length === 0) {
-      return null;
-    }
-
-    return this.ensureSubscriberVaultAnchor({
-      environmentId: params.environmentId,
-      organizationId: params.organizationId,
-      subscriberMongoId: params.subscriberMongoId,
-      agentMcpServerIds: params.agentMcpServerIds,
-      oauthEnablements,
-      runtimeProvider: params.runtimeProvider,
-    });
-  }
-
-  /**
-   * Create and persist a subscriber vault before OAuth so sessions can opt in
-   * via `vault_ids` when credentials are pushed after the in-thread Connect flow.
-   *
-   * Race-safe across concurrent first-time dispatches:
-   *   1. Re-check siblings inside the function (covers races that resolved
-   *      between the caller's read and our entry).
-   *   2. Provision a vault upstream.
-   *   3. Either propagate onto existing connection rows OR insert an anchor
-   *      row. Both write paths converge on `setIfMissing` semantics so the
-   *      first writer wins and subsequent racers read the winner's vault id.
-   *      The loser's upstream vault is logged for follow-up cleanup.
-   */
-  private async ensureSubscriberVaultAnchor(params: {
-    environmentId: string;
-    organizationId: string;
-    subscriberMongoId: string;
-    agentMcpServerIds: string[];
-    oauthEnablements: AgentMcpServerEntity[];
-    runtimeProvider: IAgentRuntimeProvider;
-  }): Promise<string> {
-    const recheck = await this.mcpConnectionRepository.findSubscriberExternalVaultId({
-      organizationId: params.organizationId,
-      environmentId: params.environmentId,
-      subscriberId: params.subscriberMongoId,
-      agentMcpServerIds: params.agentMcpServerIds,
-    });
-
-    if (recheck) {
-      return recheck;
-    }
-
-    const { externalVaultId } = await params.runtimeProvider.createVault({
-      displayName: `nv-sub-${params.subscriberMongoId}`,
-    });
-
-    const connections = await this.mcpConnectionRepository.findSubscriberConnectionsForAgent({
-      organizationId: params.organizationId,
-      environmentId: params.environmentId,
-      subscriberId: params.subscriberMongoId,
-      agentMcpServerIds: params.agentMcpServerIds,
-    });
-
-    if (connections.length > 0) {
-      await this.mcpConnectionRepository.setSubscriberExternalVaultIdIfMissing({
-        organizationId: params.organizationId,
-        environmentId: params.environmentId,
-        subscriberId: params.subscriberMongoId,
-        agentMcpServerIds: params.agentMcpServerIds,
-        externalVaultId,
-      });
-
-      const winner = await this.mcpConnectionRepository.findSubscriberExternalVaultId({
-        organizationId: params.organizationId,
-        environmentId: params.environmentId,
-        subscriberId: params.subscriberMongoId,
-        agentMcpServerIds: params.agentMcpServerIds,
-      });
-
-      if (winner && winner !== externalVaultId) {
-        this.logOrphanVault({
-          connection: connections[0],
-          ourVaultId: externalVaultId,
-          winnerVaultId: winner,
-          reason: 'subscriber_propagate_lost',
-        });
-
-        return winner;
-      }
-
-      return externalVaultId;
-    }
-
-    const anchor = params.oauthEnablements[0];
-    const catalog = MCP_SERVERS.find((entry) => entry.id === anchor.mcpId);
-
-    if (!catalog?.oauth) {
-      this.logOrphanVault({
-        ourVaultId: externalVaultId,
-        winnerVaultId: undefined,
-        reason: 'anchor_catalog_missing_oauth',
-        organizationId: params.organizationId,
-        environmentId: params.environmentId,
-        subscriberId: params.subscriberMongoId,
-      });
-
-      return externalVaultId;
-    }
-
-    try {
-      await this.mcpConnectionRepository.create({
-        _organizationId: params.organizationId,
-        _environmentId: params.environmentId,
-        scope: McpConnectionScopeEnum.Subscriber,
-        mcpId: anchor.mcpId,
-        _agentMcpServerId: anchor._id,
-        _subscriberId: params.subscriberMongoId,
-        authMode: catalog.oauth.mode,
-        status: McpConnectionStatusEnum.PendingOAuth,
-        auth: { externalVaultId },
-      });
-
-      return externalVaultId;
-    } catch (err) {
-      if (!isDuplicateKeyError(err)) {
-        throw err;
-      }
-
-      // Another writer inserted the anchor row first. Either the winner
-      // already carries a vault id (we have an orphan upstream and must use
-      // theirs) or they inserted without one (e.g. `GenerateMcpOAuthUrl`
-      // racing ahead) and we can still claim the slot.
-      const winner = await this.mcpConnectionRepository.findSubscriberExternalVaultId({
-        organizationId: params.organizationId,
-        environmentId: params.environmentId,
-        subscriberId: params.subscriberMongoId,
-        agentMcpServerIds: params.agentMcpServerIds,
-      });
-
-      if (winner) {
-        this.logOrphanVault({
-          ourVaultId: externalVaultId,
-          winnerVaultId: winner,
-          reason: 'subscriber_anchor_insert_lost',
-          organizationId: params.organizationId,
-          environmentId: params.environmentId,
-          subscriberId: params.subscriberMongoId,
-        });
-
-        return winner;
-      }
-
-      await this.mcpConnectionRepository.setSubscriberExternalVaultIdIfMissing({
-        organizationId: params.organizationId,
-        environmentId: params.environmentId,
-        subscriberId: params.subscriberMongoId,
-        agentMcpServerIds: params.agentMcpServerIds,
-        externalVaultId,
-      });
-
-      const claimed = await this.mcpConnectionRepository.findSubscriberExternalVaultId({
-        organizationId: params.organizationId,
-        environmentId: params.environmentId,
-        subscriberId: params.subscriberMongoId,
-        agentMcpServerIds: params.agentMcpServerIds,
-      });
-
-      if (claimed && claimed !== externalVaultId) {
-        this.logOrphanVault({
-          ourVaultId: externalVaultId,
-          winnerVaultId: claimed,
-          reason: 'subscriber_setif_lost',
-          organizationId: params.organizationId,
-          environmentId: params.environmentId,
-          subscriberId: params.subscriberMongoId,
-        });
-
-        return claimed;
-      }
-
-      return externalVaultId;
-    }
-  }
-
-  /**
-   * Resolve the MCP server list for a session turn, filtering to only
-   * the MCPs the subscriber can actually use right now.
-   *
-   * - Non-OAuth MCPs: always included (no per-subscriber auth needed).
-   * - OAuth MCPs: included only when the subscriber has a `Connected` row.
-   *
-   * Returns `undefined` when the agent has no OAuth MCPs (no filtering
-   * needed — the agent definition already carries the full list).
-   */
-  async resolveSubscriberConnectedMcps(params: {
-    agentId: string;
-    environmentId: string;
-    organizationId: string;
-    subscriberMongoId?: string;
-  }): Promise<{ name: string; url: string }[] | undefined> {
-    const enablements = await this.agentMcpServerRepository.findByAgent({
-      organizationId: params.organizationId,
-      environmentId: params.environmentId,
-      agentId: params.agentId,
-      enabledOnly: true,
-    });
-
-    if (enablements.length === 0) {
-      return undefined;
-    }
-
-    const oauthEnablements = enablements.filter((row) =>
-      MCP_SERVERS.some((entry) => entry.id === row.mcpId && entry.oauth)
-    );
-
-    if (oauthEnablements.length === 0) {
-      return undefined;
-    }
-
-    const connectedMcpIds = new Set<string>();
-
-    if (params.subscriberMongoId && oauthEnablements.length > 0) {
-      const connections = await this.mcpConnectionRepository.findSubscriberConnectionsForAgent({
-        organizationId: params.organizationId,
-        environmentId: params.environmentId,
-        subscriberId: params.subscriberMongoId,
-        agentMcpServerIds: oauthEnablements.map((row) => row._id),
-      });
-
-      for (const conn of connections) {
-        if (conn.status === McpConnectionStatusEnum.Connected) {
-          connectedMcpIds.add(conn.mcpId);
-        }
-      }
-    }
-
-    const result: { name: string; url: string }[] = [];
-
-    for (const enablement of enablements) {
-      const catalog = MCP_SERVERS.find((entry) => entry.id === enablement.mcpId);
-
-      if (!catalog) {
-        continue;
-      }
-
-      const isOAuth = !!catalog.oauth;
-
-      if (isOAuth && !connectedMcpIds.has(enablement.mcpId)) {
-        continue;
-      }
-
-      result.push({ name: catalog.name, url: catalog.url });
-    }
-
-    return result;
-  }
-
   private async listAgentMcpServerIds(
     params: {
       agentId: string;
@@ -478,13 +160,4 @@ export class McpConnectionVaultService {
       'Lost vault-creation race; upstream Anthropic vault was created but not adopted (orphan)'
     );
   }
-}
-
-/**
- * MongoDB duplicate-key error sentinel (E11000). Avoids importing the full
- * mongodb driver types — every Mongoose write path wraps the native error
- * with `.code === 11000`.
- */
-function isDuplicateKeyError(err: unknown): boolean {
-  return Boolean(err && typeof err === 'object' && 'code' in err && (err as { code: unknown }).code === 11000);
 }
