+++ apps/api/src/app/agents/mcp/index.ts
@@ -2,3 +2,4 @@ export { McpConnectionVaultService } from './connections/mcp-connection-vault.se
 export { AgentsMcpOAuthController } from './oauth/agents-mcp-oauth.controller';
 export { McpOAuthDiscoveryService } from './oauth/mcp-oauth-discovery.service';
 export { AgentMcpDefinitionService } from './runtime/agent-mcp-definition.service';
+export { AgentMcpSessionService } from './runtime/agent-mcp-session.service';
